선착순 3차 적용

+++ src/main/java/com/sts/couponapi/entity/FinishEvent.java
@@ -25,5 +25,5 @@ public class FinishEvent {
     private Double date;
 
     @Column(name = "count")
-    private int count;
+    private Double count;
 }
